project-5

+++ 5-project/controllers/movieController.js
@@ -10,14 +10,39 @@ const homepage = async (req, res) => {
 
 // Render form page
 const RenderForm = (req, res) => {
-    res.render("form");
+    try {
+        res.render("form");
+    } catch (error) {
+        res.send(`not found addmovie page`, error)
+    }
 };
 
 // horrer page
 const horrerPage = (req, res) => {
-    res.render('horrer');
+    try {
+        res.render('horrer');
+    } catch (error) {
+        res.send(`not found horrer page : `, error)
+    }
 }
 
+// free movie page
+const freeMoviePage = (req, res) => {
+    try {
+        res.render('freeMovie');
+    } catch (error) {
+        res.send(`not found free movie page : `, error);
+    }
+}
+
+//new movie page
+const AllMoviePage = (req, res) => {
+    try {
+        res.render('allMovie');
+    } catch (error) {
+        res.send(`<p>all movie page not found : ${error} </p>`)
+    }
+}
 // indetails page
 const inDetailsPage = async (req, res) => {
     try {
@@ -43,39 +68,40 @@ const CardInDetails = (req, res) => {
             avatar: "/images/Pokemon Season-1.jpg",
             directorName: "Kunihiko Yuyama",
             releaseYear: 1997,
-            Genre: "Action, Adventure, Comedy",
-            description: "Ash starts his journey as a Pokémon Trainer from Pallet Town with Pikachu.",
+            Genre: "Action",
+            description: "October 14, 1999 – July 27, 2000.Ash starts his journey as a Pokémon Trainer from Pallet Town with Pikachu. New Pokémon from Gen II (Gold/Silver) introduced.Starts catching Johto-native Pokémon like Chikorita and Cyndaquil.",
         },
         2: {
             movieTitle: "Pokémon: Adventures in the Orange Islands",
             avatar: "/images/Pokemon Season-2.webp",
             directorName: "Kunihiko Yuyama",
             releaseYear: 1999,
-            Genre: "Action, Adventure, Comedy",
+            Genre: " Adventure",
+            description: "August 9, 2001 – November 14, 2002.Battles tough opponents like Harrison.Misty and Brock part ways temporarily at the end of the season.Ash begins a new journey in Johto with Pikachu and friends."
         },
         3: {
             movieTitle: "Pokémon: The Johto Journeys",
             avatar: "/images/Pokemon Season-3.jpg",
             directorName: "Kunihiko Yuyama",
             releaseYear: 1999,
-            Genre: "Action, Adventure, Comedy",
-            description: "Ash begins a new journey in Johto with Pikachu and friends.",
+            Genre: "Comedy",
+            description: "October 14, 1999 – July 27, 2000.New Pokémon from Gen II (Gold/Silver) introduced. Rivalry with new characters like Casey.Builds strong bonds with his Johto Pokémon.Ash begins a new journey in Johto with Pikachu and friends.",
         },
         4: {
             movieTitle: "Pokémon: Johto League Champions",
             avatar: "/images/Pokemon Season-4.jpg",
             directorName: "Kunihiko Yuyama",
             releaseYear: 2000,
-            Genre: "Action, Adventure, Comedy",
-            description: "Continuation of Ash's Johto journey with more battles and growth.",
+            Genre: "Action, Comedy",
+            description: "August 3, 2000 – August 2, 2001.Faces more Gym Leaders and trains his Johto team. Continuation of Ash's Johto journey with more battles and growth.",
         },
         5: {
             movieTitle: "Pokémon: Master Quest",
             avatar: "/images/Pokemon Season-5.jpg",
             directorName: "Kunihiko Yuyama",
             releaseYear: 2001,
-            Genre: "Action, Adventure, Comedy",
-            description: "Final part of the Johto saga, leading up to the Silver Conference.",
+            Genre: "Action",
+            description: " January 28, 1999 – October 7, 1999.New companion: Tracey joins the group.Returns with Lapras and Charizard becoming more cooperative.Final part of the Johto saga, leading up to the Silver Conference.",
         }
     };
     try {
@@ -87,7 +113,7 @@ const CardInDetails = (req, res) => {
             res.send("Series not found");
         }
     } catch (error) {
-        res.send(`card is not found ` , error)
+        res.send(`Series is not found `, error)
     }
 };
 
@@ -114,16 +140,20 @@ const insertMovie = async (req, res) => {
 const DeletMovie = async (req, res) => {
     try {
         const id = req.params.id;
-        const record = await movie.findById(id);
+        const data = await movie.findById(id);
 
-        if (record.avatar) {
-            fs.unlinkSync(record.avatar);
+        if (data) {
+            console.log(data.avatar)
+            fs.unlinkSync(data.avatar);
+            await movie.findByIdAndDelete(id);
+            res.redirect('/');
         }
-        await movie.findByIdAndDelete(id);
-        res.redirect('/');
+        else{
+            console.log("Single Record not found");
+        }
     } catch (error) {
-        console.error("Error deleting movie card:", error);
-        res.send("Delete error");
+        console.error("Error delete movie information:", error);
+        res.send("delete error");
     }
 };
 
@@ -163,5 +193,5 @@ const EditMovie = async (req, res) => {
 };
 
 module.exports = {
-    homepage, RenderForm, horrerPage, inDetailsPage, CardInDetails, insertMovie, DeletMovie, UpdateMovie, EditMovie
+    homepage, RenderForm, horrerPage, freeMoviePage, AllMoviePage, inDetailsPage, CardInDetails, insertMovie, DeletMovie, UpdateMovie, EditMovie
 };
